+патрон в патроннике для магазинного оружия
Учитывается в режиме AF_AMMO_BOX_AS_MAGAZINE, без него это несколько бессмысленно.
При полной перезарядке и смене типа боеприпаса заряжается сначала полный магазин. Если повторить перезарядку и в запасе есть полные магазины будет дозаряжен ещё один полный магазин + сохранён патрон в патроннике.
При смене патрона разряжается полный магазин + 1 патрон.
Для этого пришлось перенести сброс флажка оружейного спавна патронов из TryAmmoToBelt в switch2_Idle (на этом этапе все разрядки-перезарядки уже точно закончены), иначе флажок сбрасывался раньше чем обрабатывался перенос патрона на пояс. Не очень красивое решение, но ничего лучше пока не придумал.
При разрядке оружия в инвентаре разряжается только магазин, патрон из патронника надо извлекать самому передёргиванием затвора.
Пока будет так, потом можно просто для контекстного меню вместо UnloadMagazine вызывать UnloadAmmo(iAmmoElapsed)

для помповых дробовиков:
+сброс зума при перезарядке
+перезарядку можно остановить кнопкой перезарядки (как-то оно прям просилось)

~вернул кодировку для сообщений об ошибках

+++ trunk/xray/xr_3da/xrGame/xrServer_Objects_ALife_Items.cpp
@@ -581,7 +581,7 @@ CSE_ALifeItemWeaponMagazined::CSE_ALifeItemWeaponMagazined	(LPCSTR caSection) : 
 	if (FireModesList) 
 	{
 		int ModesCount = _GetItemCount(FireModesList);
-		m_u8CurFireMode = ModesCount - 1;
+		m_u8CurFireMode = (u8)ModesCount - 1;
 	}
 	else 
 	{
